Extract S2933 report grouping helper
